Addition of main function and sub packages for tools
