+++ go/main.go
@@ -1,0 +1,33 @@
+/*
+Copyright 2021 The Vitess Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"flag"
+	"github.com/vitessio/arewefastyet/go/tools/microbench"
+)
+
+func main() {
+	var pkg string
+	var output string
+
+	flag.StringVar(&pkg, "in", "", "Go package to benchmark")
+	flag.StringVar(&output, "out", "", "Output file")
+	flag.Parse()
+
+	microbench.MicroBenchmark(pkg, output)
+}
